testing spam categorization
this is mostly just to have something pushed to the repo so that i can revert if i mess up too much. the main addition in this commit is finding a way to get each spam email and then actually going into one of them and grabbing some information from the label.