desktop application
in this update, i created a desktop app to allow users to interact natively and with a very simple process. next step will be to package the app and then make a basic website where users can go to download the zip and see basic instructions.

also, i will manually upload `spam_app/study_venv/bin/Activate.ps1`. not sure why it's fatal to replace CRLF with LF but it'll be fine, i'll figure it out.