complete runner.sh
i think the runner file is finished. if homebrew isn't installed, this installs it. if xterm isn't installed, this installs it. if python isn't installed, this installs it. if pip isn't installed, this installs it. if selenium or undetected_chromedriver isn't installed, this installs it. then this runs the python script in an xterm window. all that's left is to make this an executable :)