final updates to website
website should be all up to date. we'll have to push the python file again later because of the time-stamp thing but that shouldn't be too much of an issue. for now, this works. good job us :)

resolved issues:
download issue: resolved by using a zip file instead of a plaintext file
nonexistent chrome issue: resolved by showing it as a prerequisite and offering a download link.
bash permissions: fixed by making it a .command file