hide scrollbars
i just wanted to be able to hide the gd scrollbars. here's why: when you're on the home page all of the content is shifted a few pixels to the left because of the scrollbar, and then when you click into the python instructions page all the content shifts back to the right because it's not long enough for a scrollbar. without the scrollbar there's no content-shifting which makes it look nicer. at least to me.