removed instructions & fixed download
since we've fixed the 2 factor authentication issue, we don't need that part in the instructions anymore. also removed it from the README.

realized that the command file won't work on windows, we need to create a batch file which will hopefully work. in the script we added a bit that changes the link to the download file based on the os (pretty proud of that lol).