MistAke
i'm gonna uppercase another letter for every time i have to edit directly in main, if i make it to "MISTAKE" i'm quitting this project.

the issue this time was just that the DesktopNotifier constructor takes either no arguments or both arguments. idk if that's actually the issue but i'm hoping it is.