MIStAKe ( fixed :) )
make it with only 2 to go. pretty solid. basically fixed two things here.
1) first, the desktop notification. apparently the previous package we were using didn't work for windows so i decided to make a direct call and use an executable file that takes care of it in windows. in mac, the same happens it just will always run the windows thing first. when it returns an error (bc mac doesn't have a zsh or bash command called `notify-send`) it'll automatically switch over to the mac version which essentially runs an applescript bit of code to show a desktop notification. 
2) second, chrome versions. i knew this would become an issue, just didn't realize it'd be this soon. chrome released version 100 recently, probably sometime in the last 2 days. because of that, the chromedriver gets screwy with version issues. i fixed that using a try-catch statement. this will still only work assuming that the user has at least chrome 98. i think we can put it up as a prerequisite to run the download.
i'll double check that this works on mac in a bit but i'm pretty sure we're good here. if we are good then i will start cutting process.py into pieces for readability and then i'll try to put them in a folder together? idk i'll figure it out as i go.