a step in the right direction
this works to log the user in but doesn't do anything with the token yet. the next step is figuring out how to use the token to actually access gmail information for each user. i think this is a step in the right direction though, let's see where this goes.