changeup
changed the scope of the gmail api to metadata because it requires less security and offers more confidence to the user that we aren't recording anything personal. also added a few screenshots to instructions, will need to finish this with the checkboxes part once i figure out how to do that.